backend/requirements.txt
aioredis==1.3.1
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.1.1
Automat==20.2.0
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.0
constantly==15.1.0
cryptography==36.0.1
cssselect2==0.4.1
daphne==3.0.2
Deprecated==1.2.13
Django==3.2.18
django-cleanup==6.0.0
django-cors-headers==3.11.0
django-filter==21.1
django-redis==5.2.0
django-tinymce4-lite==1.8.0
djangorestframework==3.14.0
easy-thumbnails==2.8.1
exif==1.6.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
jsmin==3.0.1
lxml==4.8.0
msgpack==1.0.4
packaging==21.3
Pillow==9.0.1
plum-py==0.8.5
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.7
pytz==2021.3
redis==4.1.4
reportlab==3.6.7
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
svglib==1.2.1
tinycss2==1.1.1
Twisted==22.1.0
txaio==21.2.1
typing_extensions==4.1.1
webencodings==0.5.1
wrapt==1.13.3
zope.interface==5.4.0
